+++ authy/src/index.ts
@@ -1,38 +1,5 @@
-import express from "express";
-import "express-async-errors";
-import { json } from "body-parser";
 import mongoose from "mongoose";
-import cookieSession from "cookie-session";
-
-import { currentUserRouter } from "./routes/current-user";
-import { signinRouter } from "./routes/signin";
-import { signupRouter } from "./routes/signup";
-import { signoutRouter } from "./routes/signout";
-import { errorHandler } from "./middlewares/error-handler";
-import { NotFoundError } from "./errors/not-found-error";
-
-const app = express();
-app.set('trust proxy', true);
-app.use(json());
-app.use(
-  cookieSession({
-    signed: false,
-    secure: true
-  }) 
-);
-
-/** Routes Setup */
-app.use(currentUserRouter);
-app.use(signinRouter);
-app.use(signoutRouter);
-app.use(signupRouter);
-
-app.all("*", async (req, res) => {
-  throw new NotFoundError();
-});
-
-app.use(errorHandler);
-
+import { app } from "./app";
 /** setting up db */
 const start = async () => {
   if(!process.env.JWT_KEY){
